+++ src/motion/__tests__/custom.test.tsx
@@ -21,4 +21,22 @@ describe("motion.custom", () => {
 
         render(<Component />)
     })
+
+    test("doesn't forward motion props", () => {
+        let animate: any
+        const BaseComponent = React.forwardRef(
+            (props: Props, ref: RefObject<HTMLDivElement>) => {
+                animate = (props as any).animate
+                return <div ref={ref} />
+            }
+        )
+
+        const MotionComponent = motion.custom<Props>(BaseComponent)
+
+        const Component = () => <MotionComponent foo animate={{ x: 100 }} />
+
+        render(<Component />)
+
+        expect(animate).toBeUndefined()
+    })
 })